Move init test to its own file
