+++ src/tests/mod.rs
@@ -300,187 +300,7 @@ fn test_create_ec_objects() {
     testtokn.finalize();
 }
 
-#[test]
-fn test_init_token() {
-    let mut testtokn = TestToken::new("test_init_token.sql");
-
-    let mut args = testtokn.make_init_args();
-    let args_ptr = &mut args as *mut CK_C_INITIALIZE_ARGS;
-    let mut ret = fn_initialize(args_ptr as *mut std::ffi::c_void);
-    assert_eq!(ret, CKR_OK);
-    let mut session: CK_SESSION_HANDLE = CK_UNAVAILABLE_INFORMATION;
-    let mut ro_session: CK_SESSION_HANDLE = CK_UNAVAILABLE_INFORMATION;
-
-    /* init once */
-    let pin_value = "SO Pin Value";
-    ret = fn_init_token(
-        testtokn.get_slot(),
-        CString::new(pin_value).unwrap().into_raw() as *mut u8,
-        pin_value.len() as CK_ULONG,
-        std::ptr::null_mut(),
-    );
-    assert_eq!(ret, CKR_OK);
-
-    /* verify wrong SO PIN fails */
-    let bad_value = "SO Bad Value";
-    ret = fn_init_token(
-        testtokn.get_slot(),
-        CString::new(bad_value).unwrap().into_raw() as *mut u8,
-        pin_value.len() as CK_ULONG,
-        std::ptr::null_mut(),
-    );
-    assert_eq!(ret, CKR_PIN_INCORRECT);
-
-    /* re-init */
-    let pin_value = "SO Pin Value";
-    ret = fn_init_token(
-        testtokn.get_slot(),
-        CString::new(pin_value).unwrap().into_raw() as *mut u8,
-        pin_value.len() as CK_ULONG,
-        std::ptr::null_mut(),
-    );
-    assert_eq!(ret, CKR_OK);
-
-    /* login as so */
-    ret = fn_open_session(
-        testtokn.get_slot(),
-        CKF_SERIAL_SESSION | CKF_RW_SESSION,
-        std::ptr::null_mut(),
-        None,
-        &mut session,
-    );
-    assert_eq!(ret, CKR_OK);
-    ret = fn_login(
-        session,
-        CKU_SO,
-        CString::new(pin_value).unwrap().into_raw() as *mut u8,
-        pin_value.len() as CK_ULONG,
-    );
-    assert_eq!(ret, CKR_OK);
-
-    /* change so pin */
-    let new_pin = "New SO Pin Value";
-    ret = fn_set_pin(
-        session,
-        CString::new(pin_value).unwrap().into_raw() as *mut u8,
-        pin_value.len() as CK_ULONG,
-        CString::new(new_pin).unwrap().into_raw() as *mut u8,
-        new_pin.len() as CK_ULONG,
-    );
-    assert_eq!(ret, CKR_OK);
-
-    /* try to open ro_session and fail */
-    ret = fn_open_session(
-        testtokn.get_slot(),
-        CKF_SERIAL_SESSION,
-        std::ptr::null_mut(),
-        None,
-        &mut ro_session,
-    );
-    assert_eq!(ret, CKR_SESSION_READ_WRITE_SO_EXISTS);
-
-    /* logout and retry */
-    ret = fn_logout(session);
-    assert_eq!(ret, CKR_OK);
-    ret = fn_open_session(
-        testtokn.get_slot(),
-        CKF_SERIAL_SESSION,
-        std::ptr::null_mut(),
-        None,
-        &mut ro_session,
-    );
-    assert_eq!(ret, CKR_OK);
-
-    /* try to change pin and fail with ro_session */
-    ret = fn_set_pin(
-        ro_session,
-        CString::new(pin_value).unwrap().into_raw() as *mut u8,
-        pin_value.len() as CK_ULONG,
-        CString::new(new_pin).unwrap().into_raw() as *mut u8,
-        new_pin.len() as CK_ULONG,
-    );
-    assert_eq!(ret, CKR_SESSION_READ_ONLY);
-
-    /* try to login again and fail because of ro_session exists */
-    ret = fn_login(
-        session,
-        CKU_SO,
-        CString::new(new_pin).unwrap().into_raw() as *mut u8,
-        new_pin.len() as CK_ULONG,
-    );
-    assert_eq!(ret, CKR_SESSION_READ_ONLY_EXISTS);
-
-    /* try again after closing ro_session */
-    ret = fn_close_session(ro_session);
-    assert_eq!(ret, CKR_OK);
-    ret = fn_login(
-        session,
-        CKU_SO,
-        CString::new(new_pin).unwrap().into_raw() as *mut u8,
-        new_pin.len() as CK_ULONG,
-    );
-    assert_eq!(ret, CKR_OK);
-
-    /* set user pin */
-    let user_pin = "User PIN Value";
-    ret = fn_init_pin(
-        session,
-        CString::new(user_pin).unwrap().into_raw() as *mut u8,
-        user_pin.len() as CK_ULONG,
-    );
-    assert_eq!(ret, CKR_OK);
-
-    /* try to log in as user and fail because SO active */
-    ret = fn_login(
-        session,
-        CKU_USER,
-        CString::new(user_pin).unwrap().into_raw() as *mut u8,
-        user_pin.len() as CK_ULONG,
-    );
-    assert_eq!(ret, CKR_USER_ANOTHER_ALREADY_LOGGED_IN);
-
-    /* retry user login after logout */
-    ret = fn_logout(session);
-    assert_eq!(ret, CKR_OK);
-    ret = fn_login(
-        session,
-        CKU_USER,
-        CString::new(user_pin).unwrap().into_raw() as *mut u8,
-        user_pin.len() as CK_ULONG,
-    );
-    assert_eq!(ret, CKR_OK);
-
-    /* change user pin as user */
-    let new_user_pin = "New User PIN Value";
-    ret = fn_set_pin(
-        session,
-        CString::new(user_pin).unwrap().into_raw() as *mut u8,
-        user_pin.len() as CK_ULONG,
-        CString::new(new_user_pin).unwrap().into_raw() as *mut u8,
-        new_user_pin.len() as CK_ULONG,
-    );
-    assert_eq!(ret, CKR_OK);
-
-    ret = fn_logout(session);
-    assert_eq!(ret, CKR_OK);
-
-    /* change back user pin after logout */
-    ret = fn_set_pin(
-        session,
-        CString::new(new_user_pin).unwrap().into_raw() as *mut u8,
-        new_user_pin.len() as CK_ULONG,
-        CString::new(user_pin).unwrap().into_raw() as *mut u8,
-        user_pin.len() as CK_ULONG,
-    );
-    assert_eq!(ret, CKR_OK);
-
-    ret = fn_logout(session);
-    assert_eq!(ret, CKR_OK);
-    ret = fn_close_session(session);
-    assert_eq!(ret, CKR_OK);
-
-    testtokn.finalize();
-}
+mod init;
 
 #[test]
 fn test_get_mechs() {
